fix: correct payroll mint to 2.083X multiplier, fix domain registration auth (#2069)
Payroll mint (CBE spec §6):
- Multiplier: 1.25X → 2.083X (gross = X × 25/12, where X = collaborator amount)
- Split: 20% SOV treasury / 32% locked reserve / 48% collaborator
- Debt ceiling: 71M → 147.9M CBE (gross, equivalent to 71M collaborator-received)
- Governance guard: signer must be Bootstrap Council member (DID derived from dilithium_pk)
- SOVRN audit tracks reserve portion value

Domain registration:
- Fix fee_payment_tx ownership check: compare dilithium_pk directly instead of key_id
  (wallet registry stores only dilithium_pk, so key_id computed with zeroed kyber_pk
  mismatches the transaction signature which carries both components)

u128 widening:
- EmploymentRegistry: compensation_amount, gross/tax/net amounts, voting_power
- CreateEmploymentContractData.compensation_amount u64 → u128
- CLI --compensation arg widened to u128

lib-client key handling:
- cbe_tx builders use create_public_key_with_kyber() to include kyber_pk in signatures
  (fixes key_id mismatch for identities with real kyber keys)

CLI:
- Add --keystore flag to payroll command for council member signing

+++ lib-blockchain/src/transaction/core.rs
@@ -1512,7 +1512,7 @@ impl Transaction {
         dao_id: [u8; 32],
         employee_key_id: [u8; 32],
         contract_type: u8,
-        compensation_amount: u64,
+        compensation_amount: u128,
         payment_period: u8,
         tax_rate_basis_points: u16,
         tax_jurisdiction: String,
@@ -2146,8 +2146,8 @@ pub struct CreateEmploymentContractData {
     pub employee_key_id: [u8; 32],
     /// Contract type discriminant (0=PublicAccess, 1=Employment)
     pub contract_type: u8,
-    /// Compensation amount in CBE atomic units
-    pub compensation_amount: u64,
+    /// Compensation amount in CBE atomic units (18-decimal)
+    pub compensation_amount: u128,
     /// Payment period discriminant (0=Monthly, 1=Quarterly, 2=Annually)
     pub payment_period: u8,
     /// Tax rate in basis points (max 5000 = 50%)
@@ -2160,10 +2160,10 @@ pub struct CreateEmploymentContractData {
 
 /// Payroll mint as a synthetic CBE bonding curve event (CBE spec §6).
 ///
-/// Mints `amount_cbe` (X) to the collaborator and routes 0.25X to the SOV
-/// treasury, recording a PRE_BACKED entry for the full 1.25X gross.  No SOV
-/// enters the system — s_c does not change.  A governance-approved deliverable
-/// hash must be recorded on-chain before the mint fires (protocol invariant).
+/// Mints gross ≈ 2.083X CBE, split 20% SOV treasury / 32% locked reserve /
+/// 48% collaborator (= X).  Records a PRE_BACKED entry for the full gross.
+/// No SOV enters the system — s_c does not change.  Signer must be a
+/// Bootstrap Council member (governance guard).
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ProcessPayrollData {
     /// Employment contract that authorises this payroll (32-byte contract_id).
